backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pymysql==1.1.0
cryptography==41.0.7
django-cors-headers==4.3.0
python-decouple==3.8
drf-spectacular==0.26.5
django-filter==23.3
